Got rid of on-the-fly signature check
The tor dev signing key got huge. It takes >10min on my machine to do anything with gpg because of that key.
Therefore, Jitpack nearly always times out while either retreiving the tor dev signature key or while verifying the signature.
Now, the signature check is done by the packager and checksums are commited to the repo for jitpack to check against.